requirements.txt
fastapi[standard]
sentence-transformers
faiss-cpu
numpy
google-genai
python-dotenv
deep-translator
pydantic
google